Create common webpage header style with flex
Gave the header class a display of flex,
justify-content of space-between and
align-items of center.

The flex algorithm on treats all the direct
children of the header class as single flex items
despite the fact the left-links and and right-links
divs have uls as their direct children. This is the
reason why space-between will flush the left and right
links to the start (left) and end (right) of the main
axis.

The ul is not a direct child of the header class so
we can make it a flex container itself and give its
items a gap of 20px. Removed default margin and padding
that was affecting the left items.

Long read. Worth it.