Create a common website feature with max-width
Pay close attention to the max width property of the
card because without it, the items will fill up the
width along the primary axis.

Justify-content: space between will not be visible
because they have taken all the space already.
With the max-width: 200px applied. The text div
adjusts and the text wraps. There is thus space
to justify content.