Align the contents of a modal with flex
Pay close attention to the icon class.
if we wrap it in a div, the flex property
that we added to the modal class will treat
the icon as a flex item together with the
container div and  and it will not be smashed
by the text or rather the container div.

Another approach was to set a flex-shrink: 0
property for the icon