Align the content div to the center
Everything is quite understandable but to one
way to stick the footer at the bottom is to give
the body a flex direction of column and the footer
a margin top of auto. This works but it pushes the
content upwards - away from the center.

Another way to do this would be to the remove the
margin top auto from the footer and give the content
 a property of flex: 1. The is shorthand of
flex: 1 1 0 and the flex-grow here does the magic.
It makes the content div grow to take all
available space between the header and the footer
while respecting the height of the body (100vh).
If the body height is removed, the header, content
and footer shrinks and the footer doesn't stick to
the bottom.

If the content div was given a height of say 600px,
without the flex: 1, it will still keep the footer
at the bottom but flex is preferable because it
adjust the growth of the content based on the
user's viewport. 100vh = 100% viewport height.

Thanks to Lisa on TOP's Discord.
100vh = 100