Pay attention to the margin on card items
The solution to this recommended the property
of margin: 16px to the 6 card items. This causes
3 rows with 2 cards each (3x2). A better
solution was to wrap the card items into a
cards div and justify-content: space-evenly.
Space-evenly remove unequal space.
Giving the card a max-width (solution recommends
with) of 300px and gives a 2x3 layout (2 rows 3 columns).

Is this recommended?